Some changes based on testing ST and ADXL

+++ HipMounted_Pedometer/Core/Inc/ADC.h
@@ -14,10 +14,10 @@
 /* Externs */
 extern ADC_HandleTypeDef hadc1;
 extern const float STM_res; // STM resolution
-extern const float refV; // reference Voltage[V]
-extern const float sensitivity;// [V/g] 330 mV/g
-extern const float zero_gBias; // ADC value at 0g [max ADC value divided by 2]
-extern const float ADC_per_gVal; // The ADC value between a difference in 1 g - [(sensitivity/ref voltage)*(STM resolution)]
+extern const double refV; // reference Voltage[V]
+extern const double sensitivity;// [V/g] 330 mV/g
+extern const double zero_gBias; // ADC value at 0g [max ADC value divided by 2]
+extern const double ADC_per_gVal; // The ADC value between a difference in 1 g - [(sensitivity/ref voltage)*(STM resolution)]
 
 /* Function Declarations */
 /*
@@ -25,8 +25,8 @@ extern const float ADC_per_gVal; // The ADC value between a difference in 1 g - 
  * */
 void getValues(void);
 
-int ADC_to_V(uint32_t ADC_val);
-int g_to_ADC(float g_val);
+double ADC_to_V(uint32_t ADC_val);
+double g_to_ADC(double g_val);
 void get_ADC_Values(void);
